Added ability for any amount of CAD files to be uploaded into an SQLITE database
Successfully got all of the data normalized into an SQLITE DDL from the CAD files that were provided to me with the project requirements. The UI scaffolding looks complete and now that I have the data in a database, it's just a matter of a little extra UI tweaking to get the APIs working.